Finish Project

+++ src/br/com/alura/playlist/modelos/Audio.java
@@ -23,7 +23,7 @@ public class Audio {
         return totalCurtidas;
     }
 
-    public double getClassificacao() {
+    public int getClassificacao() {
         return classificacao;
     }
